requirements.txt
flask==2.2.2
flask-sqlalchemy==3.0.3
Werkzeug==2.2.2
sqlalchemy-serializer==1.4.1
flask-migrate==4.0.0
flask-restful==0.3.9
importlib-metadata==6.0.0
importlib-resources==5.10.0
pytest==7.2.0
python-dotenv==1.0.1
sqlalchemy==2.0.25
flask-bcrypt==1.0.1
faker==15.3.2
gunicorn==23.0.0
alembic==1.14.1
aniso8601==10.0.0
attrs==24.3.0
click==8.1.8
exceptiongroup==1.2.2
greenlet==3.1.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
Mako==1.3.8
MarkupSafe==2.1.5
numpy==1.24.4
packaging==24.2
pandas==2.0.3
pluggy==1.5.0
psycopg2==2.9.4
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.17.0
scikit-learn==1.3.2
tomli==2.2.1
typing_extensions==4.12.2
zipp==3.20.2
python-docx==1.1.0
PyPDF2==3.0.1
pdfplumber==0.10.3
flask-cors==4.0.0